Move some code.
