+++ src/output/mod.rs
@@ -1,6 +1,8 @@
 //! Types describing how to format output.
+mod cssbuf;
 mod format;
 mod style;
+mod transform;
 
 pub use format::{Format, Formatted};
 pub use style::Style;
